extract FilesCollector
